# contracts/openapi/yumpoo-v1.yaml
openapi: 3.0.3
info:
  title: Yumpoo API
  version: 1.0.0
  description: |
    Yumpoo 一期 HTTP API 的唯一静态契约源。

    所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共
    schema、header、parameter、response 和 golden example；业务 paths 在对应模块
    实现时按兼容新增原则加入。
servers:
  - url: /api/v1
    description: Yumpoo v1 business API
...
paths:
...
  /work-items/{workItemId}/updates:
    get:
      tags: [WorkItemUpdates]
      operationId: listWorkItemUpdates
      summary: 读取 Work Item 讨论窗口
      description: 仅返回未删除主评论。首窗附带全部置顶评论，普通窗口仍包含置顶记录，客户端按 ID 去重；窗口按 createdAt、id 升序，nextCursor 加载更早记录。
      security: [{ sessionCookie: [] }]
      parameters:
        - name: workItemId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: cursor
          in: query
          description: 服务端签发的不透明 Base64URL v1 向前游标。
          schema: { type: string, maxLength: 256 }
        - name: size
          in: query
          schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: 可见讨论窗口；归档 Project 或 Content 的历史仍可读取。
          headers: { Cache-Control: { schema: { type: string }, description: 固定为 no-store。 } }
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdatePage' } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      tags: [WorkItemUpdates]
      operationId: publishWorkItemUpdate
      summary: 发布 Work Item 讨论
      description: 正文中的 Mention 是唯一提及输入；服务端净化 HTML、重写权威显示名并原子发布事件。
      security: [{ sessionCookie: [] }]
      parameters:
        - name: workItemId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/XsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkItemUpdateCreateRequest' }
            examples:
              richText: { $ref: '#/components/examples/WorkItemUpdateRichText' }
      responses:
        '201':
          description: Update 已发布或精确幂等重放；父 Work Item 的版本与更新时间不变。
          headers:
            Location: { $ref: '#/components/headers/Location' }
            ETag: { $ref: '#/components/headers/ETag' }
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdate' } } }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /work-item-updates/{updateId}/pin:
    patch:
      tags: [WorkItemUpdates]
      operationId: pinWorkItemUpdate
      summary: 置顶或取消置顶主评论
      description: 可发表评论的项目成员均可操作任意主评论，子评论不可置顶；重复状态不增版。
      security: [{ sessionCookie: [] }]
      parameters:
        - name: updateId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/XsrfToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdatePinRequest' } } }
      responses:
        '200':
          description: 置顶状态已更新或没有变化。
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { schema: { type: string }, description: 固定为 no-store。 }
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdate' } } }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /work-item-updates/{updateId}/replies:
    get:
      tags: [WorkItemUpdates]
      operationId: listWorkItemUpdateReplies
      summary: 按创建时间正序分页读取未删除回复
      security: [{ sessionCookie: [] }]
      parameters:
        - name: updateId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: cursor
          in: query
          schema: { type: string, maxLength: 256 }
        - name: size
          in: query
          schema: { type: integer, format: int32, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: 正序回复窗口，pinnedItems 恒为空。
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdatePage' } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /work-item-updates/{updateId}:
    get:
      tags: [WorkItemUpdates]
      operationId: getWorkItemUpdate
      summary: 查询单条 Work Item 讨论
      description: 返回当前调用人的操作能力；删除记录仅返回占位事实，不返回原正文。
      security: [{ sessionCookie: [] }]
      parameters:
        - name: updateId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200':
          description: 当前可见 Update。
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { schema: { type: string }, description: 固定为 no-store。 }
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdate' } } }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
    patch:
      tags: [WorkItemUpdates]
      operationId: editWorkItemUpdate
      summary: 编辑 Work Item 讨论
      description: 仅作者可编辑，不限制发布时间；正文和 Mention 重新净化并原子替换。
      security: [{ sessionCookie: [] }]
      parameters:
        - name: updateId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/XsrfToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdateEditRequest' } } }
      responses:
        '200':
          description: Update 已编辑，或净化后的正文没有变化。
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { schema: { type: string }, description: 固定为 no-store。 }
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdate' } } }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
    delete:
      tags: [WorkItemUpdates]
      operationId: deleteWorkItemUpdate
      summary: 删除 Work Item 评论或讨论串
      description: 作者可不限时删除自己的评论；当前 ProjectOwner 可删除所有评论，包括归档历史。删除主评论会原子删除全部回复。
      security: [{ sessionCookie: [] }]
      parameters:
        - name: updateId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/XsrfToken'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdateDeleteRequest' } } }
      responses:
        '200':
          description: Update 已转为不可恢复的删除占位。
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Cache-Control: { schema: { type: string }, description: 固定为 no-store。 }
          content: { application/json: { schema: { $ref: '#/components/schemas/WorkItemUpdate' } } }
        '400': { $ref: '#/components/responses/MalformedRequest' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ResourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-yumpoo-session
      description: 服务端签发的 HttpOnly、Secure、SameSite=Lax Web Session Cookie。
  schemas:
...
    Uuid:
      type: string
      format: uuid
      description: 不透明 UUID；客户端不得从值中推导业务语义。
...
    WorkItemUpdateStatus:
      type: string
      enum: [PUBLISHED, EDITED, DELETED]
    WorkItemUpdate:
      type: object
      additionalProperties: false
      required: [id, projectId, contentId, workItemId, authorUserId, authorDisplayName,
        bodyHtml, bodyText, status, rowVersion, etag, createdAt,
        editedAt, editedByUserId, deletedAt, deletedByUserId, parentUpdateId, pinnedAt, pinnedByUserId, replyCount, replies, repliesNextCursor, capabilities]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        projectId: { $ref: '#/components/schemas/Uuid' }
        contentId: { $ref: '#/components/schemas/Uuid' }
        workItemId: { $ref: '#/components/schemas/Uuid' }
        authorUserId: { $ref: '#/components/schemas/Uuid' }
        authorDisplayName: { type: string, minLength: 1, maxLength: 200 }
        bodyHtml: { type: string, minLength: 1, maxLength: 65536, nullable: true }
        bodyText: { type: string, minLength: 1, maxLength: 16384, nullable: true }
        status: { $ref: '#/components/schemas/WorkItemUpdateStatus' }
        rowVersion: { type: integer, format: int64, minimum: 0, readOnly: true }
        etag: { type: string, pattern: '^"[0-9]+"$', readOnly: true }
        createdAt: { type: string, format: date-time, readOnly: true }
        editedAt: { type: string, format: date-time, nullable: true, readOnly: true }
        editedByUserId: { type: string, format: uuid, nullable: true, readOnly: true }
        deletedAt: { type: string, format: date-time, nullable: true, readOnly: true }
        deletedByUserId: { type: string, format: uuid, nullable: true, readOnly: true }
        parentUpdateId: { type: string, format: uuid, nullable: true, readOnly: true }
        pinnedAt: { type: string, format: date-time, nullable: true, readOnly: true }
        pinnedByUserId: { type: string, format: uuid, nullable: true, readOnly: true }
        replyCount: { type: integer, format: int64, minimum: 0, readOnly: true }
        replies:
          type: array
          items: { $ref: '#/components/schemas/WorkItemUpdate' }
        repliesNextCursor: { type: string, maxLength: 256, nullable: true }
        capabilities: { $ref: '#/components/schemas/WorkItemUpdateCapabilities' }
    WorkItemUpdateCapabilities:
      type: object
      additionalProperties: false
      required: [canEdit, canDelete, canReply, canPin]
      properties:
        canEdit: { type: boolean, readOnly: true }
        canDelete: { type: boolean, readOnly: true }
        canReply: { type: boolean, readOnly: true }
        canPin: { type: boolean, readOnly: true }
    WorkItemUpdatePage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, pinnedItems]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/WorkItemUpdate' }
        nextCursor: { type: string, maxLength: 256, nullable: true }
        pinnedItems:
          type: array
          items: { $ref: '#/components/schemas/WorkItemUpdate' }
    WorkItemUpdateCreateRequest:
      type: object
      additionalProperties: false
      required: [bodyHtml]
      properties:
        parentUpdateId: { type: string, format: uuid, nullable: true, description: 省略时发布主评论；指定时仅允许回复同一工作项的未删除主评论。 }
        bodyHtml:
          type: string
          description: >-
            未受信任的协作富文本；净化后 HTML 最多 65536 字符、纯文本非空且最多 16384 字符。
            允许 p/br/strong/em/u/s/h2/pre/hr/ul/ol/li/blockquote/code/a/span/table/thead/tbody/tr/th/td。
            span 仅保留受限的 color/background-color（十六进制或 0–255 整数 rgb）和 font-size（16/18/24/32/36/48px）；
            p/h2 仅允许 text-align 为 left/center/right；p/h2/pre/blockquote/li 的 dir 仅允许 ltr/rtl。
            表格单元格 rowspan/colspan 为 1–20，ol.start 为 1–10000。
            清单使用 ul[data-type=taskList] 与直接子 li[data-type=taskItem][data-checked=true|false]，已发布正文只读。
            Mention 使用 span[data-type=mention][data-mention-user-id=UUID]，由服务端校验 ACTIVE 项目成员并重写显示名。
            链接仅允许绝对 http/https/mailto；移除任意 class、事件、未知样式及图片等标签。客户端仅展示净化响应。
    WorkItemUpdateEditRequest:
      type: object
      additionalProperties: false
      required: [bodyHtml]
      properties:
        bodyHtml:
          type: string
          description: 未受信任的编辑正文；格式白名单与 WorkItemUpdateCreateRequest.bodyHtml 相同，服务端重新净化并从正文解析 Mention。
    WorkItemUpdateDeleteRequest:
      type: object
      additionalProperties: false
      properties: {}
    WorkItemUpdatePinRequest:
      type: object
      additionalProperties: false
      required: [pinned]
      properties:
        pinned: { type: boolean }
...
    RequestId:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
      description: 服务端生成或从可信反向代理接收的不透明请求关联标识。
    ErrorCode:
      type: string
      description: |
        一期冻结的公共错误码。客户端只能使用 code 做程序判断，不能解析 message。
      enum:
        - MALFORMED_REQUEST
        - AUTHENTICATION_REQUIRED
        - ACCOUNT_DISABLED
        - ACCESS_DENIED
        - RESOURCE_NOT_FOUND
        - IDEMPOTENCY_KEY_REUSED
        - REQUEST_IN_PROGRESS
        - INVALID_STATE_TRANSITION
        - WORKLOG_LOCKED
        - VERSION_CONFLICT
        - FILE_TOO_LARGE
        - FILE_TYPE_NOT_ALLOWED
        - VALIDATION_FAILED
        - CLIENT_UPGRADE_REQUIRED
        - PRECONDITION_REQUIRED
        - RATE_LIMITED
        - INTERNAL_ERROR
        - DEPENDENCY_UNAVAILABLE
    FieldError:
      type: object
      additionalProperties: false
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          minLength: 1
          description: 出错字段的稳定请求字段名。
        code:
          type: string
          minLength: 1
          description: 字段级稳定错误码。
        message:
          type: string
          minLength: 1
          description: 面向当前用户且不包含内部实现信息的说明。
    ApiErrorBase:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
        - retryable
        - fieldErrors
        - details
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
          description: 面向当前用户的安全错误说明。
        requestId:
          $ref: '#/components/schemas/RequestId'
        retryable:
          type: boolean
          description: 当前请求是否可在不改变业务意图的前提下自动重试。
        fieldErrors:
          type: array
          description: 字段级错误；非字段校验错误固定为空数组。
          items:
            $ref: '#/components/schemas/FieldError'
        details:
          type: object
          description: 错误专用详情；具体响应 schema 必须进一步收窄。
    EmptyErrorDetails:
      type: object
      additionalProperties: false
      description: |
        通用错误详情默认为空；需要稳定机器可读原因的业务冲突可返回 reason。
      properties:
        reason:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{1,63}$'
          description: M2-08 生命周期治理稳定原因；其他业务域可继续使用各自稳定原因。
          x-stable-values:
            - PROJECT_ARCHIVE_BLOCKED
            - WORKSPACE_ARCHIVE_BLOCKED
            - WORKSPACE_UNCHANGED
        blockers:
          type: array
          description: 仅包含稳定分类 code 与非负聚合 count，不含对象标识或业务正文。
          items: { $ref: '#/components/schemas/SafeBlocker' }
    SafeBlocker:
      type: object
      additionalProperties: false
      required: [code, count]
      properties:
        code:
          type: string
          enum: [OPEN_WORK_ITEMS, PENDING_WORKLOG_APPROVALS, OPEN_PRODUCT_FEEDBACK, CURRENT_PROJECTS]
        count: { type: integer, format: int64, minimum: 0 }
    ErrorResponse:
      description: ApiErrorBase 的具体响应类型；details 默认为空，特定冲突可带稳定 reason。
      allOf:
        - $ref: '#/components/schemas/ApiErrorBase'
        - type: object
          required:
            - details
          properties:
            details:
              $ref: '#/components/schemas/EmptyErrorDetails'
...
  parameters:
...
    XsrfToken:
      name: X-XSRF-TOKEN
      in: header
      required: true
      description: 与当前 Session 绑定并由非 HttpOnly Cookie 回传的 CSRF 凭据。
      schema:
        type: string
        minLength: 1
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: |
        单个强 ETag，格式为双引号包围的十进制 rowVersion。弱验证器、`*` 和多个
        ETag 均不属于一期契约。
      schema:
        type: string
        pattern: '^"[0-9]+"$'
      example: '"5"'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 客户端为单一业务意图生成并在重试时复用的 UUID。
      schema:
        type: string
        format: uuid
        maxLength: 64
...
  headers:
    Location:
      description: 跳转目标 URI。
      schema:
        type: string
        format: uri-reference
...
    XRequestId:
      description: 请求关联标识；错误响应中必须与 body.requestId 完全一致。
      schema:
        $ref: '#/components/schemas/RequestId'
    ETag:
      description: 当前资源 rowVersion 对应的单个强 ETag。
      schema:
        type: string
        pattern: '^"[0-9]+"$'
      example: '"5"'
    RetryAfter:
      description: 允许再次尝试前应等待的秒数。
      schema:
        type: integer
        format: int32
        minimum: 1
  examples:
    WorkItemUpdateRichText:
      summary: 受限富文本、表格、清单及个人提及的发布输入
      externalValue: ../examples/work-items/update-rich-text.json
...
    MalformedRequest:
      summary: JSON、类型、参数或 If-Match 格式无法解析
      externalValue: ../examples/errors/400-malformed-request.json
    AuthenticationRequired:
      summary: 会话缺失、未知、过期、主动退出或登录回调统一失败
      externalValue: ../examples/errors/401-authentication-required.json
    AccountDisabled:
      summary: 可识别会话因离职或账号禁用被撤销
      externalValue: ../examples/errors/403-account-disabled.json
    AccessDenied:
      summary: 有效主体可见资源但没有动作权限
      externalValue: ../examples/errors/403-access-denied.json
    ResourceNotFound:
      summary: 资源不存在或存在性对当前主体不可见
      externalValue: ../examples/errors/404-resource-not-found.json
    IdempotencyKeyReused:
      summary: 相同幂等键被用于不同请求内容
      externalValue: ../examples/errors/409-idempotency-key-reused.json
    RequestInProgress:
      summary: 相同幂等请求仍在处理中
      externalValue: ../examples/errors/409-request-in-progress.json
    InvalidStateTransition:
      summary: 当前业务状态不允许目标命令
      externalValue: ../examples/errors/409-invalid-state-transition.json
    WorklogLocked:
      summary: 已提交或批准工时不能普通修改
      externalValue: ../examples/errors/409-worklog-locked.json
    VersionConflict:
      summary: If-Match 与当前可见资源版本不一致
      externalValue: ../examples/errors/412-version-conflict.json
...
    ValidationFailed:
      summary: 请求格式正确但字段或业务规则不满足
      externalValue: ../examples/errors/422-validation-failed.json
...
    PreconditionRequired:
      summary: 条件写入缺少 If-Match
      externalValue: ../examples/errors/428-precondition-required.json
...
  responses:
...
    MalformedRequest:
      description: 请求 JSON、类型、参数格式或条件头无法解析。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            malformedRequest:
              $ref: '#/components/examples/MalformedRequest'
    AuthenticationRequired:
      description: 未建立有效 CurrentActor。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            authenticationRequired:
              $ref: '#/components/examples/AuthenticationRequired'
    Forbidden:
      description: 可识别账号已禁用，或有效主体缺少可见资源上的动作权限。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            accountDisabled:
              $ref: '#/components/examples/AccountDisabled'
            accessDenied:
              $ref: '#/components/examples/AccessDenied'
    ResourceNotFound:
      description: 资源不存在，或资源存在性对当前主体不可见。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            resourceNotFound:
              $ref: '#/components/examples/ResourceNotFound'
    Conflict:
      description: 幂等、处理中、状态迁移或工时锁定冲突。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            idempotencyKeyReused:
              $ref: '#/components/examples/IdempotencyKeyReused'
            requestInProgress:
              $ref: '#/components/examples/RequestInProgress'
            invalidStateTransition:
              $ref: '#/components/examples/InvalidStateTransition'
            worklogLocked:
              $ref: '#/components/examples/WorklogLocked'
    VersionConflict:
      description: If-Match 与当前可见资源版本不一致。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            versionConflict:
              $ref: '#/components/examples/VersionConflict'
...
    ValidationFailed:
      description: 请求格式正确，但字段、必填项或业务规则不满足。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationFailed:
              $ref: '#/components/examples/ValidationFailed'
...
    PreconditionRequired:
      description: 条件写入缺少必需的 If-Match。
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            preconditionRequired:
              $ref: '#/components/examples/PreconditionRequired'
...