# packages/api-client/src/generated/models/WorkItemUpdateEditRequest.ts
/* tslint:disable */
/**
 * Yumpoo API
 * Yumpoo 一期 HTTP API 的唯一静态契约源。  所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共 schema、header、parameter、response 和 golden example；业务 paths 在对应模块 实现时按兼容新增原则加入。
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface WorkItemUpdateEditRequest
 */
export interface WorkItemUpdateEditRequest {
    /**
     * 未受信任的编辑正文；格式白名单与 WorkItemUpdateCreateRequest.bodyHtml 相同，服务端重新净化并从正文解析 Mention。
     * @type {string}
     * @memberof WorkItemUpdateEditRequest
     */
    bodyHtml: string;
}

/**
 * Check if a given object implements the WorkItemUpdateEditRequest interface.
 */
export function instanceOfWorkItemUpdateEditRequest(value: object): value is WorkItemUpdateEditRequest {
    if (!('bodyHtml' in value) || value['bodyHtml'] === undefined) return false;
    return true;
}

export function WorkItemUpdateEditRequestFromJSON(json: any): WorkItemUpdateEditRequest {
    return WorkItemUpdateEditRequestFromJSONTyped(json, false);
}

export function WorkItemUpdateEditRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkItemUpdateEditRequest {
    if (json == null) {
        return json;
    }
    return {

        'bodyHtml': json['bodyHtml'],
    };
}

export function WorkItemUpdateEditRequestToJSON(json: any): WorkItemUpdateEditRequest {
    return WorkItemUpdateEditRequestToJSONTyped(json, false);
}

export function WorkItemUpdateEditRequestToJSONTyped(value?: WorkItemUpdateEditRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'bodyHtml': value['bodyHtml'],
    };
}