# packages/api-client/src/generated/models/WorkItemUpdatePinRequest.ts
/* tslint:disable */
/**
 * Yumpoo API
 * Yumpoo 一期 HTTP API 的唯一静态契约源。  所有业务 operation 都必须挂载在相对 server `/api/v1` 下。M0-09 只冻结公共 schema、header、parameter、response 和 golden example；业务 paths 在对应模块 实现时按兼容新增原则加入。
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface WorkItemUpdatePinRequest
 */
export interface WorkItemUpdatePinRequest {
    /**
     *
     * @type {boolean}
     * @memberof WorkItemUpdatePinRequest
     */
    pinned: boolean;
}

/**
 * Check if a given object implements the WorkItemUpdatePinRequest interface.
 */
export function instanceOfWorkItemUpdatePinRequest(value: object): value is WorkItemUpdatePinRequest {
    if (!('pinned' in value) || value['pinned'] === undefined) return false;
    return true;
}

export function WorkItemUpdatePinRequestFromJSON(json: any): WorkItemUpdatePinRequest {
    return WorkItemUpdatePinRequestFromJSONTyped(json, false);
}

export function WorkItemUpdatePinRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkItemUpdatePinRequest {
    if (json == null) {
        return json;
    }
    return {

        'pinned': json['pinned'],
    };
}

export function WorkItemUpdatePinRequestToJSON(json: any): WorkItemUpdatePinRequest {
    return WorkItemUpdatePinRequestToJSONTyped(json, false);
}

export function WorkItemUpdatePinRequestToJSONTyped(value?: WorkItemUpdatePinRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'pinned': value['pinned'],
    };
}